+++ controllers/userCtrl.js
@@ -5,25 +5,8 @@ const gravatar = require("gravatar");
 const path = require("path");
 const Jimp = require("jimp");
 const fs = require("fs").promises;
-const multer = require("multer");
 
 const storeImage = path.join(process.cwd(), "public", "avatars");
-
-const tempDir = path.join(__dirname, "/../../temp");
-
-const storage = multer.diskStorage({
-  destination: tempDir,
-  filename: (req, file, cb) => {
-    cb(null, file.originalname);
-  },
-  limits: {
-    fileSize: 1048576,
-  },
-});
-
-const upload = multer({
-  storage: storage,
-});
 
 const registrationUser = async (req, res, next) => {
   const { email, password: pass, subscription } = req.body;
@@ -161,5 +144,4 @@ module.exports = {
   currentUser,
   updateUserAvatar,
   uploadUserAvatar,
-  upload,
 };
